requirements.txt
streamlit==1.50.0
pandas==2.3.3
requests==2.32.5
scikit-learn==1.7.2
python-dotenv